+++ funfact/__init__.py
@@ -2,13 +2,16 @@
 # -*- coding: utf-8 -*-
 """Fun-Fact"""
 from .backend import use, active_backend, available_backends
-from .lang import index, indices, tensor, template, _0, _1, delta, pi
+from .lang import (
+    index, indices, tensor, template, pi, zeros, ones, eye
+)
 from .lang._math import *  # noqa: F401, F403
 from .model import Factorization
 from .algorithm import factorize
 from .vectorization import vectorize, view
 from .context import is_grad_on, enable_grad
 from . import initializers
+from . import conditions
 
 
 __all__ = [
@@ -19,9 +22,9 @@ __all__ = [
     'indices',
     'tensor',
     'template',
-    '_0',
-    '_1',
-    'delta',
+    'zeros',
+    'ones',
+    'eye',
     'pi',
     'Factorization',
     'factorize',
@@ -29,11 +32,12 @@ __all__ = [
     'view',
     'is_grad_on',
     'enable_grad',
-    'initializers'
+    'initializers',
+    'conditions'
 ]
 
 
-__version__ = '0.7.1'
+__version__ = '0.9.2'
 __author__ = '''Yu-Hang "Maxin" Tang, Daan Camps, Elizaveta Rebrova'''
 __maintainer__ = 'Yu-Hang "Maxin" Tang'
 __email__ = 'Tang.Maxin@gmail.com'