+++ funfact/lang/__init__.py
@@ -3,23 +3,28 @@
 import warnings
 from ._tsrex import index, indices, tensor, TsrEx
 from ._tplex import template
-from ._predefined_literal import _0, _1, delta, pi
+from ._constants import pi
+from ._special import zeros, ones, eye
 
 
 try:
-    from IPython import get_ipython
-    from IPython.display import display, HTML
+    import google.colab  # noqa: F401
+    try:
+        from IPython import get_ipython
+        from IPython.display import display, HTML
 
-    get_ipython().events.register(
-        'pre_run_cell',
-        lambda: display(HTML(
-            "<script type='text/javascript' async "
-            "src='https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.7/"
-            "MathJax.js?config=TeX-MML-AM_CHTML'></script>"
-        ))
-    )
-except Exception:
-    warnings.warn('Cannot set up MathJAX, LaTeX rendering may not work.')
+        get_ipython().events.register(
+            'pre_run_cell',
+            lambda: display(HTML(
+                "<script type='text/javascript' async "
+                "src='https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.7/"
+                "MathJax.js?config=TeX-MML-AM_CHTML'></script>"
+            ))
+        )
+    except Exception:
+        warnings.warn('Cannot set up MathJAX, LaTeX rendering may not work.')
+except ImportError:
+    pass
 
 
 __all__ = [
@@ -27,9 +32,9 @@ __all__ = [
     'indices',
     'tensor',
     'template',
-    '_0',
-    '_1',
-    'delta',
+    'zeros',
+    'ones',
+    'eye',
     'pi',
     'TsrEx'
 ]