+++ funfact/lang/_ast.py
@@ -11,35 +11,50 @@ class _ASNode:
     pass
 
 
+def primitive(precedence=None):
+    def make_primitive(f):
+        args = inspect.getfullargspec(f).args
+        p = make_dataclass(
+            f.__name__,
+            args,
+            bases=(_ASNode,)
+        )
+        p.name = property(lambda self: f.__name__)
+        if precedence is not None:
+            p.precedence = property(lambda self: precedence)
+        p.fields = property(lambda self: self.__dict__)
+        p.fields_fixed = property(lambda self: {
+            k: v for k, v in self.__dict__.items() if k in args
+        })
+        p.fields_payload = property(lambda self: {
+            k: v for k, v in self.__dict__.items() if k not in args
+        })
+        return p
+    return make_primitive
+
+
 class Primitives:
 
-    def primitive(precedence=None):
-        def make_primitive(f):
-            args = inspect.getfullargspec(f).args
-            p = make_dataclass(
-                f.__name__,
-                args,
-                bases=(_ASNode,)
-            )
-            p.name = property(lambda self: f.__name__)
-            if precedence is not None:
-                p.precedence = property(lambda self: precedence)
-            p.fields = property(lambda self: self.__dict__)
-            p.fields_fixed = property(lambda self: {
-                k: v for k, v in self.__dict__.items() if k in args
-            })
-            p.fields_payload = property(lambda self: {
-                k: v for k, v in self.__dict__.items() if k not in args
-            })
-            return p
-        return make_primitive
+    @primitive(precedence=1)
+    def abstract_index_notation(tensor: _ASNode, indices: _ASNode):
+        '''indexing a raw tensor or tensor expression: tensor[indices...].
+        To be interpreted either as tensor indexing or index renaming depending
+        on the type of the addressee.'''
+
+    @primitive(precedence=None)
+    def abstract_binary(
+        lhs: _ASNode, rhs: _ASNode, precedence: int, operator: str
+    ):
+        '''generic binary operations, to be interpreted as  matmul, kron,
+        elementwise, or Einstein based on index/indexless status and the
+        operator being used.'''
 
     @primitive(precedence=0)
     def literal(value: LiteralValue):
         '''a literal value'''
 
     @primitive(precedence=0)
-    def tensor(abstract: AbstractTensor):
+    def tensor(decl: AbstractTensor):
         '''an abstract tensor'''
 
     @primitive(precedence=0)
@@ -49,14 +64,12 @@ class Primitives:
         same index.'''
 
     @primitive(precedence=0)
-    def indices(items: Tuple[AbstractIndex]):
+    def indices(items: Tuple[index]):
         '''a tuple of indices'''
 
     @primitive(precedence=1)
-    def index_notation(indexless: _ASNode, indices: _ASNode):
-        '''indexing a raw tensor or indexless expression: expr[indices...].
-        To be interpreted either as tensor indexing or index renaming depending
-        on the type of the addressee.'''
+    def indexed_tensor(tensor: _ASNode, indices: _ASNode):
+        '''indexing a raw tensor or indexless expression: expr[indices...]'''
 
     @primitive(precedence=2)
     def call(f: str, x: _ASNode):
@@ -66,18 +79,11 @@ class Primitives:
     def neg(x: _ASNode):
         '''elementwise negation'''
 
-    @primitive(precedence=5)
-    def matmul(lhs: _ASNode, rhs: _ASNode):
-        '''indexless matrix multiplication'''
-
-    @primitive(precedence=5)
-    def kron(lhs: _ASNode, rhs: _ASNode):
-        '''indexless Kronecker product'''
-
     @primitive(precedence=None)
-    def binary(lhs: _ASNode, rhs: _ASNode, precedence: int, oper: str):
-        '''generic binary operations, to be interpreted as either elementwise
-        or Einstein based on index/indexless status.'''
+    def elem(
+        lhs: _ASNode, rhs: _ASNode, precedence: int, operator: str
+    ):
+        '''indexless elementwise operations between tensors'''
 
     @primitive(precedence=None)
     def ein(
@@ -88,10 +94,16 @@ class Primitives:
 
     @primitive(precedence=9)
     def tran(src: _ASNode, indices: _ASNode):
-        '''transposition/axis reordering'''
+        '''transpose the tensor by permuting the axes.'''
+
+    @primitive(precedence=9)
+    def abstract_dest(src: _ASNode, indices: indices):
+        '''generic destination index designation; maybe translated either
+        into transposition/axis permutation, or specify the output indices
+        of Einstein operations'''
 
     Tensorial = Union[
-        index_notation, call, pow, neg, ein
+        indexed_tensor, call, neg, ein
     ]
     Numeric = Union[Tensorial, Real]