+++ funfact/lang/interpreter/_evaluation.py
@@ -13,10 +13,16 @@ class Evaluator(ROOFInterpreter):
     def _binary_operator(reduction, pairwise, lhs, rhs, spec):
         return _einop(spec, lhs, rhs, reduction, pairwise)
 
+    def abstract_index_notation(self, tensor, indices, **kwargs):
+        raise NotImplementedError()
+
+    def abstract_binary(self, lhs, rhs, precedence, operator, **kwargs):
+        raise NotImplementedError()
+
     def literal(self, value, **kwargs):
         return ab.tensor(value.raw)
 
-    def tensor(self, abstract, data, **kwargs):
+    def tensor(self, decl, data, **kwargs):
         return data
 
     def index(self, item, bound, kron, **kwargs):
@@ -25,8 +31,8 @@ class Evaluator(ROOFInterpreter):
     def indices(self, items, **kwargs):
         return None
 
-    def index_notation(self, indexless, indices, **kwargs):
-        return indexless
+    def indexed_tensor(self, tensor, indices, **kwargs):
+        return tensor
 
     def call(self, f, x, **kwargs):
         return getattr(ab, f)(x)
@@ -34,14 +40,8 @@ class Evaluator(ROOFInterpreter):
     def neg(self, x, **kwargs):
         return -x
 
-    def matmul(self, lhs, rhs, **kwargs):
-        return lhs @ rhs
-
-    def kron(self, lhs, rhs, **kwargs):
-        return ab.kron(lhs, rhs)
-
-    def binary(self, lhs, rhs, precedence, oper, **kwargs):
-        return getattr(ab, oper)(lhs, rhs)
+    def elem(self, lhs, rhs, precedence, operator, **kwargs):
+        return getattr(ab, operator)(lhs, rhs)
 
     def ein(self, lhs, rhs, precedence, reduction, pairwise, outidx, einspec,
             **kwargs):
@@ -50,3 +50,6 @@ class Evaluator(ROOFInterpreter):
     def tran(self, src, indices, einspec, **kwargs):
         in_spec, out_spec = einspec.split('->')
         return ab.transpose(src, [in_spec.index(i) for i in out_spec])
+
+    def abstract_dest(self, src, indices, **kwargs):
+        raise NotImplementedError()