+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
+## 0.9.2 (2022-01-21)
+
+- Bug fixes and documentation updates.
+
+## 0.9.1 (2022-01-18)
+
+- Bug fixes.
+
+## 0.9 (2022-01-15)
+
+- Redesigned hybrid expression backend.
+- Penalty terms for abstract tensors
+
+## 0.8 (2022-01-07)
+
+- Allow hybrid indexless/Einstein expressions.
+- Various initializers for abstract tensors.
+
 ## 0.7.1 (2021-12-22)
 
 - Tons of new features.